+++ src/Rector/AddToFileReturnArrayByKeyRector.php
@@ -11,8 +11,8 @@ use Crmplease\Coder\Helper\ReturnStatementHelper;
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Return_;
 use Rector\Core\Rector\AbstractRector;
-use Rector\Core\RectorDefinition\CodeSample;
-use Rector\Core\RectorDefinition\RectorDefinition;
+use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
+use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
  * @author Mougrim <rinat@mougrim.ru>
@@ -66,9 +66,9 @@ class AddToFileReturnArrayByKeyRector extends AbstractRector
         return $this;
     }
 
-    public function getDefinition(): RectorDefinition
+    public function getRuleDefinition(): RuleDefinition
     {
-        return new RectorDefinition('Add to file to return array value "newValue" by "newKey" with check duplicates', [
+        return new RuleDefinition('Add to file to return array value "newValue" by "newKey" with check duplicates', [
             new CodeSample(
                 <<<'PHP'
 return [