+++ tests/Functional/AddToArrayByKeyTest.php
@@ -6,6 +6,8 @@ namespace Tests\Crmplease\Coder\Functional;
 use Crmplease\Coder\Constant;
 use Tests\Crmplease\Coder\FunctionalTestCase;
 
+require_once __DIR__ . '/../fixtures/constants.php';
+
 /**
  * @author Mougrim <rinat@mougrim.ru>
  */
@@ -89,45 +91,6 @@ class AddToArrayByKeyTest extends FunctionalTestCase
         $this->assertFixture($fixture);
     }
 
-    public function testClassConstant(): void
-    {
-        $fixture = 'classConstant';
-        $coder = $this->getCoder();
-        $coder->addToFileReturnArrayByKey(
-            $this->createFixtureFile($fixture),
-            [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\BarClass::class'),
-            'bar class'
-        );
-        $this->assertFixture($fixture);
-    }
-
-    public function testClassConstantReplace(): void
-    {
-        $fixture = 'classConstantReplace';
-        $coder = $this->getCoder();
-        $coder->addToFileReturnArrayByKey(
-            $this->createFixtureFile($fixture),
-            [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::class'),
-            'foo class replaced'
-        );
-        $this->assertFixture($fixture);
-    }
-
-    public function testClassConstantDuplicates(): void
-    {
-        $fixture = 'duplicates';
-        $coder = $this->getCoder();
-        $coder->addToFileReturnArrayByKey(
-            $this->createFixtureFile($fixture),
-            [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::class'),
-            'foo class'
-        );
-        $this->assertFixture($fixture);
-    }
-
     public function testConstant(): void
     {
         $fixture = 'constant';
@@ -135,7 +98,7 @@ class AddToArrayByKeyTest extends FunctionalTestCase
         $coder->addToFileReturnArrayByKey(
             $this->createFixtureFile($fixture),
             [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\BarClass::TEST'),
+            new Constant('\Tests\Crmplease\Coder\fixtures\BAR_TEST'),
             'bar constant'
         );
         $this->assertFixture($fixture);
@@ -148,7 +111,7 @@ class AddToArrayByKeyTest extends FunctionalTestCase
         $coder->addToFileReturnArrayByKey(
             $this->createFixtureFile($fixture),
             [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::TEST'),
+            new Constant('\Tests\Crmplease\Coder\fixtures\FOO_TEST'),
             'foo constant replaced'
         );
         $this->assertFixture($fixture);
@@ -161,8 +124,86 @@ class AddToArrayByKeyTest extends FunctionalTestCase
         $coder->addToFileReturnArrayByKey(
             $this->createFixtureFile($fixture),
             [],
-            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::TEST'),
+            new Constant('\Tests\Crmplease\Coder\fixtures\FOO_TEST'),
             'foo constant'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClass(): void
+    {
+        $fixture = 'class';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\BarClass::class'),
+            'bar class'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClassReplace(): void
+    {
+        $fixture = 'classReplace';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::class'),
+            'foo class replaced'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClassDuplicates(): void
+    {
+        $fixture = 'duplicates';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::class'),
+            'foo class'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClassConstant(): void
+    {
+        $fixture = 'classConstant';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\BarClass::TEST'),
+            'bar class constant'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClassConstantReplace(): void
+    {
+        $fixture = 'classConstantReplace';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::TEST'),
+            'foo class constant replaced'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testClassConstantDuplicates(): void
+    {
+        $fixture = 'duplicates';
+        $coder = $this->getCoder();
+        $coder->addToFileReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            [],
+            new Constant('\Tests\Crmplease\Coder\fixtures\FooClass::TEST'),
+            'foo class constant'
         );
         $this->assertFixture($fixture);
     }