+++ tests/Functional/AddToReturnArrayByKeyTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Tests\Crmplease\Coder\Functional;
 
+use Crmplease\Coder\Constant;
 use Tests\Crmplease\Coder\FunctionalTestCase;
 
 /**
@@ -10,7 +11,7 @@ use Tests\Crmplease\Coder\FunctionalTestCase;
  */
 class AddToReturnArrayByKeyTest extends FunctionalTestCase
 {
-    public function test(): void
+    public function testAddKey(): void
     {
         $fixture = 'SomeClass';
         $coder = $this->getCoder();
@@ -20,6 +21,20 @@ class AddToReturnArrayByKeyTest extends FunctionalTestCase
             ['path1'],
             'key2',
             'value2'
+        );
+        $this->assertFixture($fixture);
+    }
+
+    public function testAddSelfClassConstant(): void
+    {
+        $fixture = 'ReplaceSelfClassConstantClass';
+        $coder = $this->getCoder();
+        $coder->addToReturnArrayByKey(
+            $this->createFixtureFile($fixture),
+            'getArray',
+            ['path1'],
+            new Constant('\Tests\Crmplease\Coder\fixtures\AddToReturnArrayByKey\ReplaceSelfClassConstantClass::SELF_CONSTANT'),
+            'self constant replaced',
         );
         $this->assertFixture($fixture);
     }