+++ tests/fixtures/AddToReturnArrayByKey/ReplaceSelfClassConstantClassExpected.php
@@ -1,10 +1,12 @@
 <?php
+declare(strict_types=1);
+
+namespace Tests\Crmplease\Coder\fixtures\AddToReturnArrayByKey;
+
 /**
  * @author Mougrim <rinat@mougrim.ru>
  */
-namespace Tests\Crmplease\Coder\fixtures\AddToReturnArrayByKey;
-
-class SomeClass
+class ReplaceSelfClassConstantClass
 {
     public const SELF_CONSTANT = 'self constant value';
 
@@ -12,13 +14,12 @@ class SomeClass
     {
         return [
             'path1' => [
-                self::SELF_CONSTANT => 'self constant',
+                self::SELF_CONSTANT => 'self constant replaced',
                 'key0' => 'value0',
                 'closure' => function ($param) {
                     return $param ?: null;
                 },
                 'key1' => 'value1',
-                'key2' => 'value2',
             ],
         ];
     }